refactor: rename proptypes file and move it to article-lead-asset dir
