+++ packages/article/src/article-lead-asset/crop-proptypes.js
